fixed drawimagefaded overload to use doubles instead of ints, works better added "uberFactor" to the vision radius, adjust it to increase precision of vision radius and increase lag while you're at it, each increment is a power so at 1 its normal 1 to 1 width * height, at 2 its not twice as many calculations, its (width * height )^2 and so on, be careful and don't increment too much, its still neat to look at tho

+++ Open GL Basic Engine/source/glutFramework/glutFramework/oglTexture2D.h
@@ -30,7 +30,7 @@ public:
 	bool drawImage(int nWidth, int nHeight, int nX, int nY);
 	bool drawImageFaded(double amount);
 	bool drawImageFaded(double amount, int dWidth, int dHeight);
-	bool drawImageFaded(double amount, int dWidth, int dHeight, int nX, int nY);
+	bool drawImageFaded(double amount, double dWidth, double dHeight, double nX, double nY);
 	bool oglTexture2D::drawImageSegment(double topLeftX, double topLeftY, 
 										double topRightX, double topRightY,
 										double bottomLeftX, double bottomLeftY,
